+++ start/routes/boats.ts
@@ -8,6 +8,9 @@ router
     router.get('boats/new', [controllers.Boats, 'create']).as('boats.create')
     router.post('boats', [controllers.Boats, 'store']).as('boats.store')
     router.get('boats/:id', [controllers.Boats, 'show']).as('boats.show')
+    router
+      .get('boats/:id/maintenance-log.pdf', [controllers.MaintenanceLogPdf, 'download'])
+      .as('boats.maintenanceLog.download')
     router.get('boats/:id/edit', [controllers.Boats, 'edit']).as('boats.edit')
     router.put('boats/:id', [controllers.Boats, 'update']).as('boats.update')
     router.delete('boats/:id', [controllers.Boats, 'destroy']).as('boats.destroy')